Functions from connect moved to utils. Lodash flatMap used.
