+++ src/actions/index.js
@@ -2,4 +2,4 @@ import * as authActions from './auth'
 import * as queryActions from './query'
 
 export { authActions, queryActions }
-export default Object.assign({}, authActions, queryActions)
+export default { authActions, queryActions }
